refactor(arch): 按照 Clean Architecture 重新组织核心目录结构
- 将 `workers.py` 移至 `ui/` 目录，消除 UI 线程对纯核心业务层的污染
- 建立 `core/database/` 目录，收拢 Peewee ORM 模型与 HistoryManager 持久化仓储
- 建立 `core/engines/` 目录，归集 Sender、Monitor、Editor 等有状态业务引擎
- 修复并验证了所有跨层调用的相对导入路径 (Imports)
- 仅物理移动文件与修正包路径，未更改任何底层业务逻辑
